chore(gas_price_service): split into 2 concrete types, v0 and v1
fix: remove block height from error
