+++ crates/fuel-core/src/service/adapters/fuel_gas_price_provider/tests/producer_gas_price_tests.rs
@@ -1,7 +1,7 @@
 use crate::service::adapters::fuel_gas_price_provider::tests::build_provider;
 use fuel_core_gas_price_service::{
+    common::gas_price_algorithm::GasPriceAlgorithm,
     static_updater::StaticAlgorithm,
-    GasPriceAlgorithm,
 };
 
 #[tokio::test]
