+++ crates/fuel-core/src/service/sub_services.rs
@@ -38,23 +38,13 @@ use crate::{
         SubServices,
     },
 };
-#[allow(unused_imports)]
-use fuel_core_gas_price_service::fuel_gas_price_updater::{
-    algorithm_updater,
-    fuel_core_storage_adapter::FuelL2BlockSource,
-    Algorithm,
+use fuel_core_gas_price_service::v0::uninitialized_task::{
+    new_gas_price_service_v0,
     AlgorithmV0,
-    FuelGasPriceUpdater,
-    UpdaterMetadata,
-    V0Metadata,
 };
 use fuel_core_poa::{
     signer::SignMode,
     Trigger,
-};
-use fuel_core_services::{
-    RunnableService,
-    ServiceRunner,
 };
 use fuel_core_storage::{
     self,
@@ -79,7 +69,7 @@ pub type TxPoolSharedState = fuel_core_txpool::service::SharedState<
     P2PAdapter,
     Database,
     ExecutorAdapter,
-    FuelGasPriceProvider<Algorithm>,
+    FuelGasPriceProvider<AlgorithmV0>,
     ConsensusParametersProvider,
     SharedMemoryPool,
 >;
@@ -87,7 +77,7 @@ pub type BlockProducerService = fuel_core_producer::block_producer::Producer<
     Database,
     TxPoolAdapter,
     ExecutorAdapter,
-    FuelGasPriceProvider<Algorithm>,
+    FuelGasPriceProvider<AlgorithmV0>,
     ConsensusParametersProvider,
 >;
 
@@ -199,7 +189,7 @@ pub fn init_sub_services(
     let settings = consensus_parameters_provider.clone();
     let block_stream = importer_adapter.events_shared_result();
 
-    let gas_price_init = algorithm_updater::InitializeTask::new(
+    let gas_price_service_v0 = new_gas_price_service_v0(
         config.clone().into(),
         genesis_block_height,
         settings,
@@ -207,10 +197,9 @@ pub fn init_sub_services(
         database.gas_price().clone(),
         database.on_chain().clone(),
     )?;
-    let next_algo = gas_price_init.shared_data();
-    let gas_price_service = ServiceRunner::new(gas_price_init);
 
-    let gas_price_provider = FuelGasPriceProvider::new(next_algo);
+    let gas_price_provider =
+        FuelGasPriceProvider::new(gas_price_service_v0.shared.clone());
     let txpool = fuel_core_txpool::new_service(
         config.txpool.clone(),
         database.on_chain().clone(),
@@ -345,7 +334,7 @@ pub fn init_sub_services(
     #[allow(unused_mut)]
     // `FuelService` starts and shutdowns all sub-services in the `services` order
     let mut services: SubServices = vec![
-        Box::new(gas_price_service),
+        Box::new(gas_price_service_v0),
         Box::new(txpool),
         Box::new(consensus_parameters_provider_service),
     ];
