+++ crates/services/gas_price_service/src/common.rs
@@ -1,0 +1,5 @@
+pub mod fuel_core_storage_adapter;
+pub mod gas_price_algorithm;
+pub mod l2_block_source;
+pub mod updater_metadata;
+pub mod utils;
