+++ crates/services/gas_price_service/src/common/fuel_core_storage_adapter.rs
@@ -1,20 +1,18 @@
-use crate::fuel_gas_price_updater::{
+use crate::common::utils::{
     BlockInfo,
     Error as GasPriceError,
-    Error,
-    L2BlockSource,
-    Result,
     Result as GasPriceResult,
-    UpdaterMetadata,
 };
 use anyhow::anyhow;
-use fuel_core_services::stream::BoxStream;
 use fuel_core_types::fuel_types::BlockHeight;
 
 use crate::{
-    fuel_gas_price_updater::fuel_core_storage_adapter::storage::{
-        GasPriceColumn,
-        GasPriceMetadata,
+    common::{
+        fuel_core_storage_adapter::storage::{
+            GasPriceColumn,
+            GasPriceMetadata,
+        },
+        updater_metadata::UpdaterMetadata,
     },
     ports::MetadataStorage,
 };
@@ -40,16 +38,11 @@ use fuel_core_types::{
         },
         Transaction,
     },
-    services::block_importer::SharedImportResult,
 };
 use std::cmp::min;
-use tokio_stream::StreamExt;
 
 #[cfg(test)]
 mod metadata_tests;
-
-#[cfg(test)]
-mod l2_source_tests;
 
 pub mod storage;
 
@@ -61,50 +54,31 @@ where
     fn get_metadata(
         &self,
         block_height: &BlockHeight,
-    ) -> Result<Option<UpdaterMetadata>> {
+    ) -> GasPriceResult<Option<UpdaterMetadata>> {
         let metadata = self
             .storage::<GasPriceMetadata>()
             .get(block_height)
-            .map_err(|err| Error::CouldNotFetchMetadata {
+            .map_err(|err| GasPriceError::CouldNotFetchMetadata {
                 source_error: err.into(),
             })?;
         Ok(metadata.map(|inner| inner.into_owned()))
     }
 
-    fn set_metadata(&mut self, metadata: &UpdaterMetadata) -> Result<()> {
+    fn set_metadata(&mut self, metadata: &UpdaterMetadata) -> GasPriceResult<()> {
         let block_height = metadata.l2_block_height();
         let mut tx = self.write_transaction();
         tx.storage_as_mut::<GasPriceMetadata>()
             .insert(&block_height, metadata)
-            .map_err(|err| Error::CouldNotSetMetadata {
+            .map_err(|err| GasPriceError::CouldNotSetMetadata {
                 block_height,
                 source_error: err.into(),
             })?;
-        tx.commit().map_err(|err| Error::CouldNotSetMetadata {
-            block_height,
-            source_error: err.into(),
-        })?;
+        tx.commit()
+            .map_err(|err| GasPriceError::CouldNotSetMetadata {
+                block_height,
+                source_error: err.into(),
+            })?;
         Ok(())
-    }
-}
-
-pub struct FuelL2BlockSource<Settings> {
-    genesis_block_height: BlockHeight,
-    gas_price_settings: Settings,
-    committed_block_stream: BoxStream<SharedImportResult>,
-}
-
-impl<Settings> FuelL2BlockSource<Settings> {
-    pub fn new(
-        genesis_block_height: BlockHeight,
-        gas_price_settings: Settings,
-        committed_block_stream: BoxStream<SharedImportResult>,
-    ) -> Self {
-        Self {
-            genesis_block_height,
-            gas_price_settings,
-            committed_block_stream,
-        }
     }
 }
 
@@ -117,7 +91,7 @@ pub trait GasPriceSettingsProvider: Send + Sync + Clone {
     fn settings(
         &self,
         param_version: &ConsensusParametersVersion,
-    ) -> Result<GasPriceSettings>;
+    ) -> GasPriceResult<GasPriceSettings>;
 }
 
 pub fn get_block_info(
@@ -126,9 +100,8 @@ pub fn get_block_info(
     block_gas_limit: u64,
 ) -> GasPriceResult<BlockInfo> {
     let (fee, gas_price) = mint_values(block)?;
-    let height = *block.header().height();
     let used_gas =
-        block_used_gas(height, fee, gas_price, gas_price_factor, block_gas_limit)?;
+        block_used_gas(fee, gas_price, gas_price_factor, block_gas_limit)?;
     let info = BlockInfo::Block {
         height: (*block.header().height()).into(),
         gas_used: used_gas,
@@ -143,13 +116,11 @@ fn mint_values(block: &Block<Transaction>) -> GasPriceResult<(u64, u64)> {
         .last()
         .and_then(|tx| tx.as_mint())
         .ok_or(GasPriceError::CouldNotFetchL2Block {
-            block_height: *block.header().height(),
             source_error: anyhow!("Block has no mint transaction"),
         })?;
     Ok((*mint.mint_amount(), *mint.gas_price()))
 }
 fn block_used_gas(
-    block_height: BlockHeight,
     fee: u64,
     gas_price: u64,
     gas_price_factor: u64,
@@ -158,7 +129,6 @@ fn block_used_gas(
     let scaled_fee =
         fee.checked_mul(gas_price_factor)
             .ok_or(GasPriceError::CouldNotFetchL2Block {
-                block_height,
                 source_error: anyhow!(
                     "Failed to scale fee by gas price factor, overflow"
                 ),
@@ -167,42 +137,4 @@ fn block_used_gas(
     let approximate = scaled_fee.checked_div(gas_price).unwrap_or(max_used_gas);
     let used_gas = min(approximate, max_used_gas);
     Ok(used_gas)
-}
-
-#[async_trait::async_trait]
-impl<Settings> L2BlockSource for FuelL2BlockSource<Settings>
-where
-    Settings: GasPriceSettingsProvider + Send + Sync,
-{
-    async fn get_l2_block(&mut self, height: BlockHeight) -> GasPriceResult<BlockInfo> {
-        let block = &self
-            .committed_block_stream
-            .next()
-            .await
-            .ok_or({
-                GasPriceError::CouldNotFetchL2Block {
-                    block_height: height,
-                    source_error: anyhow!("No committed block found"),
-                }
-            })?
-            .sealed_block
-            .entity;
-
-        match block.header().height().cmp(&self.genesis_block_height) {
-            std::cmp::Ordering::Less => Err(GasPriceError::CouldNotFetchL2Block {
-                block_height: height,
-                source_error: anyhow!("Block precedes expected genesis block height"),
-            }),
-            std::cmp::Ordering::Equal => Ok(BlockInfo::GenesisBlock),
-            std::cmp::Ordering::Greater => {
-                let param_version = block.header().consensus_parameters_version;
-
-                let GasPriceSettings {
-                    gas_price_factor,
-                    block_gas_limit,
-                } = self.gas_price_settings.settings(&param_version)?;
-                get_block_info(block, gas_price_factor, block_gas_limit)
-            }
-        }
-    }
 }
