+++ crates/services/gas_price_service/src/v0.rs
@@ -1,0 +1,6 @@
+pub mod algorithm;
+pub mod metadata;
+pub mod service;
+#[cfg(test)]
+mod tests;
+pub mod uninitialized_task;
