+++ crates/services/gas_price_service/src/v1/algorithm.rs
@@ -1,0 +1,13 @@
+use crate::common::gas_price_algorithm::GasPriceAlgorithm;
+use fuel_core_types::fuel_types::BlockHeight;
+use fuel_gas_price_algorithm::v1::AlgorithmV1;
+
+impl GasPriceAlgorithm for AlgorithmV1 {
+    fn next_gas_price(&self) -> u64 {
+        self.calculate()
+    }
+
+    fn worst_case_gas_price(&self, block_height: BlockHeight) -> u64 {
+        self.worst_case(block_height.into())
+    }
+}
